modify: dto 타입 수정

+++ findex/src/main/java/com/example/findex/domain/Sync_Job_Log/dto/SyncJobLogDto.java
@@ -9,7 +9,7 @@ public class SyncJobLogDto {
 
     private Long id;
     private String jobType;
-    private int indexInfoId;
+    private Long indexInfoId;
     private String targetDate;
     private String worker;
     private String jobTime;
